requirements.txt
Click==7.0
dnspython==2.1.0
matplotlib==3.4.2
